requirements/base.txt
# v0.0.1

django>3.0
Pillow
libsass
easy-thumbnails
django-compressor
django-sass-processor
